refactor(api): use consistent types from @zoonk/ai instead of manual duplicates
Export schema-derived item types from AI package and replace all manually
defined duplicates in workflow files. Delete types.ts, colocating types
where they're produced.
