+++ apps/api/src/workflows/course-generation/steps/add-alternative-titles-step.ts
@@ -1,6 +1,6 @@
 import { addAlternativeTitles } from "@zoonk/core/alternative-titles/add";
 import { streamError, streamStatus } from "../stream-status";
-import { type CourseContext } from "../types";
+import { type CourseContext } from "./initialize-course-step";
 
 export async function addAlternativeTitlesStep(input: {
   course: CourseContext;
