+++ apps/api/src/workflows/course-generation/steps/generate-alternative-titles-step.ts
@@ -1,7 +1,7 @@
 import { generateAlternativeTitles } from "@zoonk/ai/tasks/courses/alternative-titles";
 import { safeAsync } from "@zoonk/utils/error";
 import { streamError, streamStatus } from "../stream-status";
-import { type CourseContext } from "../types";
+import { type CourseContext } from "./initialize-course-step";
 
 export async function generateAlternativeTitlesStep(course: CourseContext): Promise<string[]> {
   "use step";
