+++ apps/api/src/workflows/course-generation/steps/initialize-course-step.ts
@@ -1,12 +1,20 @@
-import { prisma } from "@zoonk/db";
+import { type CourseSuggestion, prisma } from "@zoonk/db";
 import { AI_ORG_SLUG } from "@zoonk/utils/constants";
 import { safeAsync } from "@zoonk/utils/error";
 import { ensureLocaleSuffix, normalizeString, toSlug } from "@zoonk/utils/string";
 import { streamError, streamStatus } from "../stream-status";
-import { type CourseContext, type CourseSuggestionData } from "../types";
+
+export type CourseContext = {
+  courseId: number;
+  courseSlug: string;
+  courseTitle: string;
+  language: string;
+  organizationId: number;
+  targetLanguage: string | null;
+};
 
 export async function initializeCourseStep(input: {
-  suggestion: CourseSuggestionData;
+  suggestion: CourseSuggestion;
   workflowRunId: string;
 }): Promise<CourseContext> {
   "use step";
