--- apps/api/src/workflows/course-generation/types.ts
@@ -1,34 +1,0 @@
-import { type GenerationStatus } from "@zoonk/db";
-
-export type ExistingCourseContent = {
-  description: string | null;
-  imageUrl: string | null;
-  hasAlternativeTitles: boolean;
-  hasCategories: boolean;
-  hasChapters: boolean;
-};
-
-export type CourseSuggestionData = {
-  id: number;
-  language: string;
-  slug: string;
-  title: string;
-  description: string;
-  targetLanguage: string | null;
-  generationStatus: GenerationStatus;
-  generationRunId: string | null;
-};
-
-export type CourseContext = {
-  courseId: number;
-  courseSlug: string;
-  courseTitle: string;
-  language: string;
-  organizationId: number;
-  targetLanguage: string | null;
-};
-
-export type GeneratedChapter = {
-  title: string;
-  description: string;
-};
